+++ common/game.h
@@ -20,7 +20,7 @@ public:
   Game(int width, int height) : map_width(width), map_height(height) {}
   bool addPlayer(const std::string &name);
   void movePlayer(const std::string &name, int x, int y, float deltaTime);
-  void updatePlayerPosition(const std::string &name, int x, int y);
+  void updatePlayerPosition(const std::string &name, float x, float y);
   std::vector<Entity> getState();
   bool isRunning();
   void stop();
